src/test_requirements.txt
-r requirements.txt
pytest-sanic==1.0.0
tox
flake8==3.6.0
pytest
aioresponses==0.6.0
aiounittest==1.2.1